requirements.txt
chromadb==0.5.5
fastapi==0.112.1
ffmpeg==1.4
gTTS==2.5.3
langchain-community==0.2.12
langchain-huggingface==0.0.3
langchain-core==0.2.33
pydub==0.25.1
requests==2.32.3
streamlit==1.42.0
uvicorn==0.30.6
PyMuPDF==1.24.9
python-dotenv==1.0.1
python-multipart==0.0.9